// src/models/Pokemon.tsx
export interface Pokemon{
  abilities: Array<any>;
  base_experience: number;
  forms: Array<any>;
  game_indices: Array<any>;
  height: number;
  held_items: Array<any>;
  id: number;
  is_defaullt: boolean;
  location_area_encounters: string;
  moves: Array<any>;
  name: string;
  order: number;
  species: Array<any>;
  sprites: Array<any>;
  stats: Array<any>;
  types: Array<any>;
  weight: number;
}